Extended the CLI interface
To support interactions with the new exit node control module, I've added a new subset of commands directly in the CLI interface, so we can isolate the responsability of parsing process args and anything related to that in this module, and not in the lib module. And I've also added the new Nodes module, responsible of dealing with Exit nodes configuration in the Tor service.